+++ BefunRep/Test/DummyResultTester.cs
@@ -3,7 +3,7 @@ namespace BefunRep.Test
 {
 	public class DummyResultTester : ResultTester
 	{
-		public override bool test(string code, int result, out string error)
+		public override bool test(string code, long result, out string error)
 		{
 			error = "DUMMY";
 			return true;
